various bug fixes and removal of obsolete code
Ensure transaction write dos not fail on writing an issue with no asset
just return error code.

Remove the block.Read (since it can panic) and replaces calls by block.Get.

Some fixes to ensure that the miner RPC JSON is correct.

+++ block/genesis.go
@@ -6,7 +6,7 @@ package block
 
 // the starting block number
 const (
-	GenesisBlockNumber = 1
+	GenesisBlockNumber = uint64(1)
 )
 
 // this is block 1, the Genesis Block
